Chess grandmaster Judit Polgr, the strongest female chess player of all time
The N queens puzzle is the challenge of placing N non-attacking queens on an NN chessboard. Write a program that solves the N queens problem.

Usage: nqueens N
If the user called the program with the wrong number of arguments, print Usage: nqueens N, followed by a new line, and exit with the status 1
where N must be an integer greater or equal to 4
If N is not an integer, print N must be a number, followed by a new line, and exit with the status 1
If N is smaller than 4, print N must be at least 4, followed by a new line, and exit with the status 1
The program should print every possible solution to the problem
One solution per line
Format: see example
You committ have to print the solutions in a specific orderdon
You are only allowed to import the sys module
Read: Queen, Backtracking